+++ docs/HISTORY.rst
@@ -2,7 +2,21 @@
 History
 =======
 
-0.0.1 (2022-03-24)
--------------------
-* Set up loading of standard data types *.tif, *.am, *.mrc, *.rec for 2D and 3D
+0.1.0 (2022-XX-XX)
+------------------
+* Build tests for all modules
 * Integrated slcpy, spindletorch and graphformer
+    * SLCPY:
+        * Loading of data types: *.tif, *.am, *.mrc, *.rec for 2D and 3D
+        * Include all slcpy modules
+        * Move Amira file output of point cloud from graphformer
+        * SetUp workflows for data pre- and post-processing 
+
+    * SPINDLETORCH
+        *
+    * GRAPHFORMER
+        *
+
+0.0.1 (2022-03-24)
+------------------
+* Initial commit